# feat(build): use swc-loader to speed up the build by 20%