--- a/docusaurus.config.js
+++ b/docusaurus.config.js
@@ -235,6 +235,23 @@ const config = {
       async: true,
     },
   ],
+  webpack: {
+    jsLoader: (isServer) => ({
+      loader: require.resolve("swc-loader"),
+      options: {
+        jsc: {
+          parser: {
+            syntax: "typescript",
+            tsx: true,
+          },
+          target: "es2017",
+        },
+        module: {
+          type: isServer ? "commonjs" : "es6",
+        },
+      },
+    }),
+  },
 };
 
 async function createConfig() {