--- a/package.json
+++ b/package.json
@@ -26,6 +26,7 @@
     "@mdx-js/react": "^1.6.22",
     "@mui/icons-material": "^5.5.0",
     "@mui/material": "^5.5.0",
+    "@swc/core": "^1.3.72",
     "clsx": "^1.1.1",
     "d3": "^7.3.0",
     "docusaurus-plugin-image-zoom": "^1.0.1",
@@ -33,7 +34,8 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-tiny-popover": "^7.2.0",
-    "react-toastify": "^8.2.0"
+    "react-toastify": "^8.2.0",
+    "swc-loader": "^0.2.3"
   },
   "browserslist": {
     "production": [