--- a/package.json
+++ b/package.json
@@ -28,7 +28,6 @@
     "@mui/material": "^5.5.0",
     "@swc/core": "^1.3.72",
     "clsx": "^1.1.1",
-    "d3": "^7.3.0",
     "docusaurus-plugin-image-zoom": "^1.0.1",
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.1",