--- a/src/pages/landing/index.module.css
+++ b/src/pages/landing/index.module.css
@@ -1,4 +1,0 @@
-/**
- * CSS files with the .module.css suffix will be treated as CSS modules
- * and scoped locally.
- */