… src/components/HomePageWave/data.js
… src/components/HomePageWave/index.js
… src/components/HomeSectionPicture/index.js
… src/components/HomepageBanner/index.js
… src/components/HomepageFeature/index.js
… src/components/HomepageIntro/index.js
… src/pages/landing/RisingWaveText.js
… src/pages/landing/TextSVGBlack.js
… src/pages/landing/index.js
… yarn.lock